requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
pyotp==2.7.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
smartapi-python==1.3.0
urllib3==1.26.13
websocket-client==1.4.2
